Test for glibc compatibility before downloading linux binary. (#3)
The binary releases are statically linked except for glibc, they require
version 2.18 or later. This means we're not compatible with RHEL/Centos 6/7.
Rather than offer the download and have it mysteriously fail, treat this
like no package for the platform (but with a different message).

We detect this by running ldd --version and parsing the output (as ldd is
part of glibc). If this fails or isn't glibc, we continue with the install.